--- a/pkg/scheduler/framework/plugins/volumebinding/BUILD
+++ b/pkg/scheduler/framework/plugins/volumebinding/BUILD
@@ -36,8 +36,15 @@ go_test(
     srcs = ["volume_binding_test.go"],
     embed = [":go_default_library"],
     deps = [
+        "//pkg/controller/volume/persistentvolume/util:go_default_library",
         "//pkg/controller/volume/scheduling:go_default_library",
+        "//pkg/scheduler/apis/config:go_default_library",
         "//pkg/scheduler/framework/v1alpha1:go_default_library",
         "//staging/src/k8s.io/api/core/v1:go_default_library",
+        "//staging/src/k8s.io/api/storage/v1:go_default_library",
+        "//staging/src/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
+        "//staging/src/k8s.io/client-go/informers:go_default_library",
+        "//staging/src/k8s.io/client-go/kubernetes/fake:go_default_library",
+        "//vendor/k8s.io/utils/pointer:go_default_library",
     ],
 )