@ahg-g 2020-06-11
/approve

This is nice, thanks! do we have a benchmark that evaluates this optimization?