@cofyc 2020-06-12
rebased on the master and fixed a UT, @msau42 can you LGTM again?