@cofyc 2020-06-12
> /approve
> 
> This is nice, thanks! do we have a benchmark that evaluates this optimization?

there is a benchmark `BenchmarkSchedulingWaitForFirstConsumerPVs` in scheduler_perf pkg:

before:

```
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/test/integration/scheduler_perf
BenchmarkSchedulingWaitForFirstConsumerPVs/500Nodes/500Pods-12         	    1000	  76636331 ns/op
PASS
ok  	k8s.io/kubernetes/test/integration/scheduler_perf	100.857s
```

 after

```
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/test/integration/scheduler_perf
BenchmarkSchedulingWaitForFirstConsumerPVs/500Nodes/500Pods-12         	    1000	  75067825 ns/op
PASS
ok  	k8s.io/kubernetes/test/integration/scheduler_perf	99.305s
```

This benchmarks the dynamic provisioning scenario with a mock storage provisioner. Another scenario is the static provisioning. Maybe we can add a bench later.