@msau42 2020-06-12
This fix would mostly improve the Immediate binding unbound case, which I don't think there's a benchmark for