@cofyc 2020-06-12
added a TODO for volume scheduling benchmarks in https://github.com/kubernetes/kubernetes/issues/90962

it's complicated to bench volume scheduling as the number of claims used in the pod and static PVs pre-provisioned in the cluster also affects the performance of scheduling (except the binding phase, as it mostly waits for PV controller and external provisioner).